remove database test class to test_database madule
