+++ test_scrap.py
@@ -1,5 +1,4 @@
 from scrap import Post
-from database import DataBase
 from conftest import post as test_db
 import requests
 import datetime
@@ -22,12 +21,3 @@ class TestPost:
         test = Post('ign', 10)
         posts = test.scrape_full_post()
         assert posts
-
-
-
-class TestDataBase:
-
-    def test_connection(self):
-        drive = DataBase()
-        verify = drive.check_connection()
-        assert verify is None
